Fix static constructors in defclass
The logic that was in the static constructor was moved to an instance
method that can be called from the Awake method. This was done to
avoid Unity's inability to guarantee the order that static constructors
will be called in. The new method is named "CarlylesMouse" on the
principled insistence of Tims Gardner who, while striking the arm
of Ramsey Nasser repeatedly, made the case that "No, it is not a
Static Constructor. It is a mouse. It belongs to Carlyle. It's
CarlylesMouse." His wisdom stands.