Remove slow path from clojure.lang.RT/classForName
Lookups for unqualified, unimported types will throw an exception (as
they should). This patch removes lookup logic that was in place due to a
mono bug where types with slashes in their names could not be found. The
compiler no longer emits such types, so the slow path is no longer
needed.